+++ packages/engine/src/wwa_audio/index.ts
@@ -1,0 +1,5 @@
+import WWAAudio from './WWAAudio';
+import WWAAudioElement from './WWAAudioElement';
+import WWAWebAudio from './WWAWebAudio';
+
+export { WWAAudio, WWAAudioElement, WWAWebAudio };
